requirements.txt
fastapi
uvicorn
gunicorn
sqlalchemy
email-validator
alembic
bcrypt
passlib
python-jose
python-multipart
